Split workspace model sidebar mutations
