+++ Tests/QuillCodeParityTests/ParityWorkspaceSidebarGateTests.swift
@@ -4,6 +4,7 @@ final class ParityWorkspaceSidebarGateTests: QuillCodeParityTestCase {
     func testWorkspaceModelDelegatesSidebarSelectionTransitions() throws {
         let modelText = try Self.appSourceText(named: "WorkspaceModel.swift")
         let threadExtensionText = try Self.appSourceText(named: "WorkspaceModelThreads.swift")
+        let sidebarModelText = try Self.appSourceText(named: "WorkspaceModelSidebar.swift")
         let threadMutationText = try Self.appSourceText(named: "WorkspaceModelThreadMutation.swift")
         let selectionText = try Self.appSourceText(named: "WorkspaceSidebarSelectionEngine.swift")
         let bulkPlannerText = try Self.appSourceText(named: "WorkspaceSidebarBulkActionPlanner.swift")
@@ -15,21 +16,24 @@ final class ParityWorkspaceSidebarGateTests: QuillCodeParityTestCase {
         XCTAssertTrue(selectionText.contains("static func selectAll"), "Select-all behavior should be directly testable.")
         XCTAssertTrue(selectionText.contains("static func toggle"), "Selection toggles should be directly testable.")
         XCTAssertTrue(selectionText.contains("static func resolve"), "Stale-ID pruning and sidebar ordering should be directly testable.")
-        XCTAssertTrue(threadExtensionText.contains("WorkspaceSidebarSelectionEngine.start"), "WorkspaceModel thread/sidebar extension should delegate selection start.")
-        XCTAssertTrue(threadExtensionText.contains("WorkspaceSidebarSelectionEngine.selectAll"), "WorkspaceModel thread/sidebar extension should delegate select-all.")
-        XCTAssertTrue(threadExtensionText.contains("WorkspaceSidebarSelectionEngine.toggle"), "WorkspaceModel thread/sidebar extension should delegate selection toggles.")
-        XCTAssertTrue(threadExtensionText.contains("setSidebarFilter"), "WorkspaceModel thread/sidebar extension should own the saved-filter transition API.")
-        XCTAssertTrue(threadExtensionText.contains("clearSidebarSelection()"), "Changing saved filters should clear hidden bulk selections.")
+        XCTAssertTrue(sidebarModelText.contains("WorkspaceSidebarSelectionEngine.start"), "WorkspaceModel sidebar extension should delegate selection start.")
+        XCTAssertTrue(sidebarModelText.contains("WorkspaceSidebarSelectionEngine.selectAll"), "WorkspaceModel sidebar extension should delegate select-all.")
+        XCTAssertTrue(sidebarModelText.contains("WorkspaceSidebarSelectionEngine.toggle"), "WorkspaceModel sidebar extension should delegate selection toggles.")
+        XCTAssertTrue(sidebarModelText.contains("setSidebarFilter"), "WorkspaceModel sidebar extension should own the saved-filter transition API.")
+        XCTAssertTrue(sidebarModelText.contains("clearSidebarSelection()"), "Changing saved filters should clear hidden bulk selections.")
         XCTAssertTrue(threadMutationText.contains("WorkspaceSidebarSelectionEngine.resolve"), "WorkspaceModel thread mutation extension should delegate stale-ID pruning and ordering.")
         XCTAssertTrue(threadMutationText.contains("func filteredSidebarItems"), "Sidebar filtering should be shared by command availability and selection resolution.")
         XCTAssertTrue(threadMutationText.contains("sidebarFilter.includes"), "Sidebar selection resolution should use the active saved filter.")
         XCTAssertTrue(bulkPlannerText.contains("struct WorkspaceSidebarBulkActionPlanner"), "Sidebar bulk action planning should live in a focused planner.")
         XCTAssertTrue(bulkPlannerText.contains("static func plan"), "Sidebar bulk action plans should be directly testable.")
         XCTAssertTrue(bulkPlannerText.contains("enum FollowUpSelection"), "Bulk action selection follow-up policy should be explicit.")
-        XCTAssertTrue(threadExtensionText.contains("WorkspaceSidebarBulkActionPlanner.plan"), "WorkspaceModel thread/sidebar extension should delegate bulk action target planning.")
+        XCTAssertTrue(sidebarModelText.contains("WorkspaceSidebarBulkActionPlanner.plan"), "WorkspaceModel sidebar extension should delegate bulk action target planning.")
         XCTAssertTrue(bulkExecutorText.contains("struct WorkspaceSidebarBulkActionExecutor"), "Sidebar bulk action execution should live in a focused executor.")
         XCTAssertTrue(bulkExecutorText.contains("static func execute"), "Sidebar bulk mutations should be directly testable.")
-        XCTAssertTrue(threadExtensionText.contains("WorkspaceSidebarBulkActionExecutor.execute"), "WorkspaceModel thread/sidebar extension should delegate bulk action execution.")
+        XCTAssertTrue(sidebarModelText.contains("WorkspaceSidebarBulkActionExecutor.execute"), "WorkspaceModel sidebar extension should delegate bulk action execution.")
+        XCTAssertFalse(threadExtensionText.contains("public func startSidebarSelection"), "Thread lifecycle extension should not own sidebar-selection API bodies.")
+        XCTAssertFalse(threadExtensionText.contains("public func performSidebarBulkAction"), "Thread lifecycle extension should not own sidebar bulk-action API bodies.")
+        XCTAssertFalse(threadExtensionText.contains("setSidebarFilter"), "Thread lifecycle extension should not own saved-filter transition API.")
         XCTAssertFalse(modelText.contains("public func startSidebarSelection"), "WorkspaceModel.swift should not own sidebar-selection API bodies.")
         XCTAssertFalse(modelText.contains("public func performSidebarBulkAction"), "WorkspaceModel.swift should not own sidebar bulk-action API bodies.")
         XCTAssertFalse(modelText.contains("public struct SidebarSelectionState"), "WorkspaceModel should not own sidebar selection state.")
